# backend/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CashX API
  version: 1.0.0
  description: Partner platform API. All money in kopecks (int64), rates in basis points.
servers:
  - url: /api/v1
paths:
...
  /cabinet/summary:
    get:
      operationId: cabinetSummary
      summary: Partner dashboard summary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummaryResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /cabinet/offers:
    get:
      operationId: cabinetOffersList
      summary: Offer catalog with partner access state
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CabinetOffer'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /cabinet/offers/{offerId}/join:
    post:
      operationId: cabinetOfferJoin
      summary: Join an active offer, creating a tracking link
      parameters:
        - $ref: '#/components/parameters/OfferId'
      responses:
        '201':
          description: Joined
          content:
            application/json:
              schema:
                type: object
                properties:
                  offer_id:
                    type: string
                  rate_bps:
                    type: integer
                  tracking_url:
                    type: string
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /cabinet/offers/{offerId}:
    get:
      operationId: cabinetOfferStats
      summary: Offer statistics (joined offers only)
      parameters:
        - $ref: '#/components/parameters/OfferId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferStatsResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
  /cabinet/payouts/config:
    get:
      operationId: cabinetPayoutsConfig
      summary: Payout configuration
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayoutRules'
        '401':
          $ref: '#/components/responses/Error'
  /cabinet/payouts:
    get:
      operationId: cabinetPayoutsList
      summary: Balance, withdrawal requests and ledger history
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    $ref: '#/components/schemas/Balance'
                  requests:
                    type: array
                    items:
                      $ref: '#/components/schemas/WithdrawalRequest'
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/LedgerEntry'
        '401':
          $ref: '#/components/responses/Error'
  /cabinet/payouts/requests:
    post:
      operationId: cabinetPayoutRequest
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [method, amount_kopecks, requisites]
              properties:
                method:
                  type: string
                  enum: [usdt, sbp]
                amount_kopecks:
                  type: integer
                  format: int64
                requisites:
                  type: string
                bank:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalRequest'
        '400':
          $ref: '#/components/responses/Error'
        '402':
          $ref: '#/components/responses/Error'
  /cabinet/payouts/requests/{id}/cancel:
    post:
      operationId: cabinetPayoutCancel
      summary: Cancel own pending withdrawal request
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalRequest'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /cabinet/referrals:
    get:
      operationId: cabinetReferrals
      summary: Referral program info
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferralsResponse'
        '401':
          $ref: '#/components/responses/Error'
  /cabinet/notifications:
    get:
      operationId: cabinetNotificationsList
      summary: Notifications with unread counter
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  unread_count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/NotificationItem'
        '401':
          $ref: '#/components/responses/Error'
  /cabinet/notifications/read-all:
    post:
      operationId: cabinetNotificationsReadAll
      summary: Mark all notifications read
      responses:
        '204':
          description: No content
        '401':
          $ref: '#/components/responses/Error'
  /cabinet/notifications/{type}/{id}/read:
    post:
      operationId: cabinetNotificationRead
      summary: Mark a single notification read
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [announcement, personal]
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
        '401':
          $ref: '#/components/responses/Error'
  /cabinet/profile:
    put:
      operationId: cabinetProfileUpdate
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                telegram_user_id:
                  type: integer
                  format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserMe'
        '401':
          $ref: '#/components/responses/Error'
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
    OfferId:
      name: offerId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Balance:
      type: object
      properties:
        available_kopecks:
          type: integer
          format: int64
        reserved_kopecks:
          type: integer
          format: int64
...
    SummaryResponse:
      type: object
      properties:
        balance:
          $ref: '#/components/schemas/Balance'
        income:
          type: object
          properties:
            today_kopecks:
              type: integer
              format: int64
            week_kopecks:
              type: integer
              format: int64
            month_kopecks:
              type: integer
              format: int64
            all_kopecks:
              type: integer
              format: int64
        funnel:
          type: object
          properties:
            clicks:
              type: integer
            unique_clicks:
              type: integer
            registrations:
              type: integer
            first_payments:
              type: integer
            income_kopecks:
              type: integer
              format: int64
        chart:
          type: array
          items:
            $ref: '#/components/schemas/DayStats'
        active_offers:
          type: array
          items:
            type: object
            properties:
              offer_id:
                type: string
              name:
                type: string
              rate_bps:
                type: integer
              tracking_url:
                type: string
              clicks:
                type: integer
              registrations:
                type: integer
              income_kopecks:
                type: integer
                format: int64
        revshare_percent_bps:
          type: integer
        telegram:
          type: object
          properties:
            connected:
              type: boolean
    CabinetOffer:
      type: object
      properties:
        offer_id:
          type: string
        project_id:
          type: string
        project_name:
          type: string
        project_logo_url:
          type: string
          nullable: true
        name:
          type: string
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, available, pending, coming_soon]
        my_rate_bps:
          type: integer
          nullable: true
        epc:
          type: number
          nullable: true
        cr:
          type: number
          nullable: true
        my_tracking_url:
          type: string
          nullable: true
...
    OfferStatsResponse:
      type: object
      properties:
        offer:
          $ref: '#/components/schemas/OfferCard'
        tracking_url:
          type: string
        summary:
          type: object
          properties:
            today:
              $ref: '#/components/schemas/OfferAgg'
            week:
              $ref: '#/components/schemas/OfferAgg'
            month:
              $ref: '#/components/schemas/OfferAgg'
            all:
              $ref: '#/components/schemas/OfferAgg'
        chart:
          type: array
          items:
            $ref: '#/components/schemas/DayStats'
        history:
          type: array
          items:
            $ref: '#/components/schemas/HistoryItem'
...
    LedgerEntry:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        amount_kopecks:
          type: integer
          format: int64
        balance_after_kopecks:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
    PayoutRules:
      type: object
      properties:
        min_withdraw_kopecks:
          type: integer
          format: int64
        usdt_rate:
          type: number
        sbp_fee_flat_kopecks:
          type: integer
          format: int64
        sbp_fee_percent_bps:
          type: integer
        updated_at:
          type: string
          format: date-time
...